\name{EMundir-package}
\alias{EMundir-package}
\alias{EMundir}
\docType{package}
\title{
\packageTitle{EMundir}
}
\description{
\packageDescription{EMundir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMundir}
\packageIndices{EMundir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EMundir}

Maintainer: \packageMaintainer{EMundir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
